+++ backend/app/services/post_services/post_service.rb
@@ -13,23 +13,22 @@ module PostServices
       { success: false, error: "Post not found", status: :not_found }
     end
 
-    def create_post(current_user, post_params)
+    def create_post(post_params)
       post = Post.new(post_params)
-      post.user = current_user
-      Rails.logger.debug("\n Post: #{post.inspect} \n")
+      post.user = @current_user
       return { success: false, error: "Failed to create post", status: :bad_request } unless post.save
       { success: true, data: PostSerializer.new(post).serializable_hash, status: :created }
     end
 
-    def update_post(current_user, post, post_params)
-      return { success: false, error: "Unauthorized to update post", status: :unauthorized } unless post.user == current_user
+    def update_post(post, post_params)
+      return { success: false, error: "Unauthorized to update post", status: :unauthorized } unless post.user == @current_user || @current_user.admin?
       return { success: false, error: "Failed to update post", messages: post.errors.full_messages, status: :bad_request } unless post.update(post_params)
       { success: true, data: PostSerializer.new(post).serializable_hash, status: :ok }
     end
 
-    def delete_post(current_user, post)
+    def delete_post(post)
       return post_not_found_error if post.nil?
-      return not_the_same_user_error unless post.user_id == current_user.id
+      return not_the_same_user_error unless post.user_id == @current_user.id || @current_user.admin?
       post.comments.each(&:destroy)
       post.destroy
       { success: true, data: "Post deleted successfully", status: :no_content }
